private-agent/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
chromadb==0.4.18
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
pypdf==3.17.4
python-docx==1.1.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1